requirements.txt
aiogram==3.*
openai==1.*
python-dotenv
aiohttp